requirements.txt
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0
matplotlib==3.9.0
numpy==1.26.4
pillow==10.3.0
streamlit==1.36.0
